Expose struct_extensions as a proc-macro
